Add reusable test class and use to test basic admin function across Sal.
